+++ app/http/middleware/AuthenticationMiddleware.py
@@ -1,5 +1,4 @@
 ''' Class for the authentication middleware '''
-from masonite.facades.Auth import Auth
 
 class AuthenticationMiddleware(object):
     ''' Middleware class which loads the current user into the request '''
@@ -9,13 +8,8 @@ class AuthenticationMiddleware(object):
 
     def before(self):
         ''' Register as a before middleware to be ran before the request '''
-        self.load_user(self.request)
-        return self.request
+        if not self.request.user():
+            self.request.redirectTo('login')
 
     def after(self):
-        ''' Register as an after middleware to be ran after the request '''
         pass
-
-    def load_user(self, request):
-        ''' Load user into the request '''
-        request.set_user(Auth(request).user())
